refactor: extract common utilities from hover.ts (closes #346)
Extract reusable Pike identifier utilities into separate module:
- isPikeIdentifierStart()
- isPikeIdentifierChar()
- getWordRangeAtPosition()

These are now available in features/utils/pike-identifier.ts for reuse
across other navigation features like definition and references.
